Updates to command descriptions, cleaning up code
